image.cpp translation unit
